Create README.md
I received a Master Degree in Health Informatics in 2013 and have yet to have an acutal position titled Informatics or IT.  I am an RN Case Manager and want so badly to work with Informatics.  I only get hired in Case Management Departments.  I am seeking the opportunity to enter.  Entry level is expected.   
This is a new platform of navigation for me.  Odd, using GitHub without formally knowing about it.  I have a lot to learn.   As much I know about clinical data and Case Management, I am not offeed a position in informatics or IT.  I still don't understand it, but I will.  Any suggestion, help, or constructive criticism is most welcomed.